:sparkler: module change for entity adapter
