+++ libs/pages/src/lib/+state/pages.selector.ts
@@ -44,7 +44,8 @@ export const selectTechSources = createSelector(
 
 export const { selectRouteParams } = getRouterSelectors();
 
-// without entity state and adapter in the reducer
+// without entity state and adapter in the reducer, here we directly get the blogs state,
+// in the other we get a dictionary data structure where we have to tap into entity for blogs and then find by key(id)
 // export const selectTechBlogById = createSelector(
 //   selectRouteParams,
 //   selectTechState,
@@ -55,5 +56,7 @@ export const { selectRouteParams } = getRouterSelectors();
 export const selectTechBlogById= createSelector(
   selectBlogsEntities,
   selectRouteParams,
-  ({blogEntities}, { id }) => blogEntities?.blogs.find(i => i.id === id)
+  // blogEntities will return a dictionary of type tech interface,
+  // we need values but finding through id wont work for tech since we have have to find it into blogs
+  ({ blogEntities}, { id }) => blogEntities?.blogs.find(i => i.id === id)
 );
