--- libs/pages/src/lib/blog/blog-type.ts
@@ -1,9 +1,0 @@
-export interface Blog {
-  id: number;
-  image: string;
-  heading: string;
-  subHeading: string;
-  blogDate: string;
-  blogDetail: string;
-  link: string;
-}
