+++ libs/pages/src/lib/model/blog-type.ts
@@ -1,0 +1,9 @@
+export interface Blog {
+  id: number;
+  image: string;
+  heading: string;
+  subHeading: string;
+  blogDate: string;
+  blogDetail: string;
+  link: string;
+}
