+++ libs/pages/src/lib/model/tech.model.ts
@@ -1,0 +1,10 @@
+import {Blog} from "./blog-type";
+import {Sources} from "./source-type";
+
+export interface Tech {
+  showTechDesc: boolean;
+  loading: boolean;
+  errorMessage: string;
+  blogs: Blog[];
+  sources: Sources[];
+}
